Rename agents to assistants in UI
