+++ libs/ui/middleware.ts
@@ -11,7 +11,7 @@ export async function middleware(req: NextRequest) {
 
   // if user is signed in and the current path is / redirect the user to /agents
   if (user && req.nextUrl.pathname === "/") {
-    return NextResponse.redirect(new URL("/agents", req.url))
+    return NextResponse.redirect(new URL("/assistants", req.url))
   }
 
   if (user) {
@@ -37,7 +37,7 @@ export async function middleware(req: NextRequest) {
 export const config = {
   matcher: [
     "/",
-    "/agents/:path*",
+    "/assistants/:path*",
     "/settings/:path*",
     "/apis/:path*",
     "/datasources/:path*",
